pyproject.toml
[project]
name = "crossref-data-extraction"
version = "0.1.0"
description = "Extract and unify data from academic papers using Crossref API"
requires-python = ">=3.9"
dependencies = [
    "requests>=2.31.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "PyPDF2>=3.0.0",
    "tenacity>=8.2.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]